+++ spyne/protocol/dictdoc.py
@@ -320,7 +320,11 @@ class SimpleDictDocument(DictDocument):
         if issubclass(inst_class, AnyDict):
             return doc
 
-        assert issubclass(inst_class, ComplexModelBase), "Patches are welcome"
+        if not issubclass(inst_class, ComplexModelBase):
+            raise NotImplementedError("Interestingly, deserializing non complex"
+                                      " types is not yet implemented. You can"
+                                      " use a ComplexModel to wrap that field."
+                                      " Otherwise, patches are welcome.")
 
         # this is for validating cls.Attributes.{min,max}_occurs
         frequencies = defaultdict(lambda: defaultdict(int))
@@ -365,11 +369,11 @@ class SimpleDictDocument(DictDocument):
                         native_v2 = v2
                     else:
                         native_v2 = self.from_string(member.type, v2,
-                                                   self.default_binary_encoding)
+                                                           self.binary_encoding)
 
                 elif issubclass(member.type, ByteArray):
                     native_v2 = self.from_string(member.type, v2,
-                                                   self.default_binary_encoding)
+                                                           self.binary_encoding)
                 else:
                     try:
                         native_v2 = self.from_string(member.type, v2)
@@ -638,8 +642,7 @@ class HierDictDocument(DictDocument):
                 raise ValidationError((key, value))
 
             if issubclass(class_, (ByteArray, File)):
-                retval = self.from_string(class_, value,
-                                                   self.default_binary_encoding)
+                retval = self.from_string(class_, value, self.binary_encoding)
             else:
                 retval = self.from_string(class_, value)
 
@@ -779,7 +782,7 @@ class HierDictDocument(DictDocument):
             return retval
 
         if issubclass(cls, (ByteArray, File)):
-            return self.to_string(cls, value, self.default_binary_encoding)
+            return self.to_string(cls, value, self.binary_encoding)
 
         return self.to_string(cls, value)