+++ spyne/protocol/soap/soap11.py
@@ -163,10 +163,8 @@ class Soap11(XmlDocument):
     type = set(XmlDocument.type)
     type.update(('soap', 'soap11'))
 
-    def __init__(self, app=None, validator=None, xml_declaration=True,
-                cleanup_namespaces=True, encoding='UTF-8', pretty_print=False):
-        super(Soap11, self).__init__(app, validator, xml_declaration,
-                                    cleanup_namespaces, encoding, pretty_print)
+    def __init__(self, *args, **kwargs):
+        super(Soap11, self).__init__(*args, **kwargs)
 
         # SOAP requires DateTime strings to be in iso format. The following
         # lines make sure custom datetime formatting via DateTime(format="...")
@@ -237,12 +235,15 @@ class Soap11(XmlDocument):
                 body_class = ctx.descriptor.out_message
 
             # decode header objects
+            #  header elements are returned in header_class order which need not match the incoming XML
             if (ctx.in_header_doc is not None and header_class is not None):
                 headers = [None] * len(header_class)
-                for i, (header_doc, head_class) in enumerate(
-                                          zip(ctx.in_header_doc, header_class)):
+                in_header_dict = dict( [(element.tag, element) for element in ctx.in_header_doc])
+                for i, head_class in enumerate(header_class):
                     if i < len(header_class):
-                        headers[i] = self.from_element(ctx, head_class, header_doc)
+                        header_doc = in_header_dict.get("{%s}%s" % (head_class.__namespace__, head_class.__type_name__), None)
+                        if header_doc is not None:
+                            headers[i] = self.from_element(ctx, head_class, header_doc)
 
                 if len(headers) == 1:
                     ctx.in_header = headers[0]